local_requirements.txt
# Requirements for running the local test script
# Install with: pip install -r local_requirements.txt

torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
pillow>=10.0.0
rembg>=2.0.50
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
loguru>=0.7.0
pydantic-settings>=2.0.0
